Include fa-solid-900.ttf as a content file
The `<None Remove="Fonts\fa-solid-900.ttf" />` line was added to remove the `fa-solid-900.ttf` font file from being treated as a "None" item in the project.

A new `<ItemGroup>` was added to include the `fa-solid-900.ttf` font file as a "Content" item, ensuring it is included in the build output directory.

The `<CopyToOutputDirectory>Always</CopyToOutputDirectory>` directive was added to ensure the font file is always copied to the output directory during the build process.

These changes ensure the font file is available at runtime and avoid conflicts or redundancy.